Extract HTTP interactions to separate module
Step toward making these interactions mocked and testable
