+++ lib/forcex.ex
@@ -1,55 +1,17 @@
 defmodule Forcex do
-  use HTTPoison.Base
-  require Logger
+  @moduledoc """
+  The main module for interacting with Salesforce
+  """
 
-  @user_agent [{"User-agent", "forcex"}]
-  @accept [{"Accept", "application/json"}]
-  @accept_encoding [{"Accept-Encoding", "gzip,deflate"}]
+  require Logger
 
   @type client :: map
   @type response :: map | {number, any}
   @type method :: :get | :put | :post | :patch | :delete
 
-  @spec process_request_headers(list({String.t, String.t})) :: list({String.t, String.t})
-  def process_request_headers(headers), do: headers ++ @user_agent ++ @accept ++ @accept_encoding
-
-  @spec process_headers(list({String.t, String.t})) :: map
-  def process_headers(headers), do: Map.new(headers)
-
-  @spec process_response(HTTPoison.Response.t) :: response
-  def process_response(%HTTPoison.Response{body: body, headers: %{"Content-Encoding" => "gzip"} = headers } = resp) do
-    %{resp | body: :zlib.gunzip(body), headers: Map.drop(headers, ["Content-Encoding"])}
-    |> process_response
-  end
-  def process_response(%HTTPoison.Response{body: body, headers: %{"Content-Encoding" => "deflate"} = headers } = resp) do
-    zstream = :zlib.open
-    :ok = :zlib.inflateInit(zstream, -15)
-    uncompressed_data = :zlib.inflate(zstream, body) |> Enum.join
-    :zlib.inflateEnd(zstream)
-    :zlib.close(zstream)
-    %{resp | body: uncompressed_data, headers: Map.drop(headers, ["Content-Encoding"])}
-    |> process_response
-  end
-  def process_response(%HTTPoison.Response{body: body, headers: %{"Content-Type" => "application/json" <> _} = headers} = resp) do
-    %{resp | body: Poison.decode!(body, keys: :atoms), headers: Map.drop(headers, ["Content-Type"])}
-    |> process_response
-  end
-  def process_response(%HTTPoison.Response{body: body, status_code: 200}), do: body
-  def process_response(%HTTPoison.Response{body: body, status_code: status}), do: {status, body}
-
-  @spec extra_options :: list
-  defp extra_options() do
-    Application.get_env(:forcex, :request_options, [])
-  end
-
   @spec json_request(method, String.t, map | String.t, list, list) :: response
   def json_request(method, url, body, headers, options) do
-    raw_request(method, url, Poison.encode!(body), headers, options)
-  end
-
-  @spec raw_request(method, String.t, map | String.t, list, list) :: response
-  def raw_request(method, url, body, headers, options) do
-    request!(method, url, body, headers, extra_options() ++ options) |> process_response
+    Forcex.Api.Http.raw_request(method, url, Poison.encode!(body), headers, options)
   end
 
   @spec post(String.t, map | String.t, client) :: response
@@ -67,7 +29,7 @@ defmodule Forcex do
   @spec delete(String.t, client) :: response
   def delete(path, client) do
     url = client.endpoint <> path
-    raw_request(:delete, url, "", client.authorization_header, [])
+    Forcex.Api.Http.raw_request(:delete, url, "", client.authorization_header, [])
   end
 
   @spec get(String.t, map | String.t, list, client) :: response
